Changed data type for Pancake order choosing

+++ environments/PancakePuzzle.cpp
@@ -245,21 +245,21 @@ bool PancakePuzzle::Path_Check(PancakePuzzleState start, PancakePuzzleState goal
 	return false;
 }
 
-std::vector<unsigned> PancakePuzzle::Get_Puzzle_Order(int order_num, unsigned num_pancakes) {
+std::vector<unsigned> PancakePuzzle::Get_Puzzle_Order(int64_t order_num, unsigned num_pancakes) {
 	std::vector<unsigned> ops;
 	assert(order_num >= 0);
 	assert(num_pancakes > 0);
 
 
-	std::vector<int> op_nums(num_pancakes -1);
+	std::vector<int64_t> op_nums(num_pancakes -1);
 
-	int num_left = 1;
-	for(int x = num_pancakes - 2; x >= 0; x--) {
+	int64_t num_left = 1;
+	for(int64_t x = num_pancakes - 2; x >= 0; x--) {
 		op_nums[x] = order_num % num_left;
 		order_num /= num_left;
 		num_left++;
 
-		for(int y = x+1; y < ((int)num_pancakes)-1; y++) {
+		for(int64_t y = x+1; y < num_pancakes-1; y++) {
 			if(op_nums[y] >= op_nums[x]) {
 				op_nums[y]++;
 			}
